+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Log the system order on the dispatcher (#414, **@RiscadoA**).
+- Fixed Time Step plugin (#989, **@joaomanita**).
 
 ### Fixed